requirements.txt
flask
python-dotenv
tinydb
bcrypt
PyJWT
